fix: use shared Zustand store for unit preference so toggle updates all components
